TSC Type Checking
Allows for easily type-checking the project with the command line. For everything to be fully type-safe, the errors have to go away :)

In some cases, things will run just fine, depending on the context of what the code does, but it is more safe to handle the cases that the code is showing an issue for. Say like you are getting a block, and the function for that returns `Block | null`, either you should account for that null case, and throw an error, or coalesce it's type that it will be safe in that case. Say like when you loop through an array with `array[i]` inside a `for` loop, it doesn't know that `i` is specific to the length of the array, so it is okay to index into the array, because it will only ever be as long as the array actually is. So for cases like that, it's okay to assert/coalesce the value access, because you know more than it does in that context.