Added JSDoc Type-Safety, Fixed Buffer overflowing, Fixed remove function not updating offset (#16)
* JSDoc Type-Safety

Found a few cases that aren't yet handled. They are mostly nullish-specific ones, that was some of the most complex things I had to handle when working on making my codebases type-safe too, before I was using a tsconfig.

* Engine Minimum Typings

We can probably change this to ES2023, depending on the versions this is meant to run on. When running on my game, the typings for this should only go up to ES2020, so that's why I changed this to that, just in case.

* Type Adjustments

Just a few changes to make things more explicit.

* TSConfig Updates

Just making things more strict in terms of type-checking. It didn't add any new errors :)

* TSC Type Checking

Allows for easily type-checking the project with the command line. For everything to be fully type-safe, the errors have to go away :)

In some cases, things will run just fine, depending on the context of what the code does, but it is more safe to handle the cases that the code is showing an issue for. Say like you are getting a block, and the function for that returns `Block | null`, either you should account for that null case, and throw an error, or coalesce it's type that it will be safe in that case. Say like when you loop through an array with `array[i]` inside a `for` loop, it doesn't know that `i` is specific to the length of the array, so it is okay to index into the array, because it will only ever be as long as the array actually is. So for cases like that, it's okay to assert/coalesce the value access, because you know more than it does in that context.

* Upgraded tsconfig to ES2022/ES2023

* Fixed type errors and added in type checks for some edge cases

* Fixed Buffer Overflowing, Fixed Remove function not updating offset

* Adjusted how the Buffer Overflow check is done

* Fixed Buffer Overflow error happening before the buffer is full

* Changed version to 1.1.2